requirements.txt
numpy==1.26.4
pandas==2.2.1
pyyaml==6.0.1
rdkit==2023.9.6
torch==2.1.2

--find-links https://data.dgl.ai/wheels/torch-2.1/cu121/repo.html
dgl